Moved substate listener of store to its own file
