+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/model/GpsMathTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3888d7bbbe390a04b9134175a683c467
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
